Test.Inspection.Core

Description

This module implements some analyses of Core expressions necessary for Test.Inspection. Normally, users of this package can ignore this module.

Documentation

slice :: [(Var, CoreExpr)] -> Var -> Slice

Selects those bindings that define the given variable (with this variable first)

pprSlice :: Slice -> SDoc

Pretty-print a slice

pprSliceDifference :: Slice -> Slice -> SDoc

Pretty-print two slices, after removing variables occurring in both

eqSlice :: Equivalence -> Slice -> Slice -> Bool

This is a heuristic, which only works if both slices have auxiliary variables in the right order. (This is mostly to work-around the buggy CSE in GHC-8.0) It also breaks if there is shadowing.

eqSlice' :: Equivalence -> Slice -> Slice -> Either SDoc ()

Like eqSlice but also returns an explanation of inequality.

freeOfType :: Slice -> [Name] -> Maybe (Var, CoreExpr)

Returns True if the given core expression mentions no type constructor anywhere that has the given name.

freeOfTerm :: Slice -> [Name] -> Maybe (Var, CoreExpr)

Returns True if the given core expression mentions no term variable anywhere that has the given name.

doesNotAllocate :: Slice -> Maybe (Var, CoreExpr)

True if the given variable binding does not allocate, if called fully satisfied.

It currently does not look through function calls, which of course could allocate. It should probably at least look through local function calls.

The variable is important to know the arity of the function.

doesNotContainTypeClasses :: Slice -> [Name] -> Maybe (Var, CoreExpr, [TyCon])