Test.Inspection

Description

This module supports the accompanying GHC plugin Test.Inspection.Plugin and adds to GHC the ability to do inspection testing.

Synopsis

To use inspection testing, you need to

enable the TemplateHaskell language extension

declare your proof obligations using inspect or inspectTest

An example module is

{-# LANGUAGE TemplateHaskell #-}
module Simple where

import Test.Inspection
import Data.Maybe

lhs, rhs :: (a -> b) -> Maybe a -> Bool
lhs f x = isNothing (fmap f x)
rhs f Nothing = True
rhs f (Just _) = False

inspect $ 'lhs === 'rhs

On GHC < 8.4, you have to explicitly load the plugin: {-# OPTIONS_GHC -fplugin=Test.Inspection.Plugin #-}

Registering obligations

inspect :: Obligation -> Q [Dec]

As seen in the example above, the entry point to inspection testing is the inspect function, to which you pass an Obligation. It will report test failures at compile time.

Since: 0.1

inspectTest :: Obligation -> Q Exp

This is a variant of inspect that allows compilation to succeed in any case, and instead indicates the result as a value of type Result, which allows seamless integration into test frameworks.

This variant ignores the expectFail field of the obligation. Instead, it is expected that you use the corresponding functionality in your test framework (e.g. tasty-expected-failure)

Since: 0.2

data Result

The result of inspectTest, which has a more or less helpful text message

Since: 0.2

Constructors

Failure String

Success String

Defining obligations

data Obligation

This data type describes an inspection testing obligation.

It is recommended to build it using mkObligation, for backwards compatibility when new fields are added. You can also use the more mnemonic convenience functions like (===) or hasNoType.

The obligation needs to be passed to inspect or inspectTest.

Since: 0.1

Constructors

Obligation

Fields

target :: Name

The target of a test obligation; invariably the name of a local definition. To get the name of a function foo, write 'foo. This requires {-# LANGUAGE TemplateHaskell #-}.

property :: Property

The property of the target to be checked.

testName :: Maybe String

An optional name for the test

expectFail :: Bool

Do we expect this property to fail? (Only used by inspect, not by inspectTest)

srcLoc :: Maybe Loc

The source location where this obligation is defined. This is filled in by inspect.

storeResult :: Maybe String

If this is Nothing, then report errors during compilation. Otherwise, update the top-level definition with this name.

Since: 0.2

mkObligation :: Name -> Property -> Obligation

Creates an inspection obligation for the given function name with default values for the optional fields.

Since: 0.1

data Equivalence

Equivalence of terms.

Since: 0.5

Constructors

StrictEquiv
strict term equality

IgnoreTypesAndTicksEquiv
ignore types and hpc ticks during the comparison

UnorderedLetsEquiv
allow permuted let bindings, ignore types and hpc tick during comparison

data Property

Properties of the obligation target to be checked.

Since: 0.1

Constructors

EqualTo Name Equivalence
Are the two functions equal?

More precisely: f is equal to g if either the definition of f is f = g, or the definition of g is g = f, or if the definitions are f = e and g = e.

In general f and g need to be defined in this module, so that their actual defintions can be inspected.

The Equivalence indicates how strict to check for equality

NoTypes [Name]
Do none of these types appear anywhere in the definition of the function (neither locally bound nor passed as arguments)

Since: 0.3

NoAllocation
Does this function perform no heap allocations.

NoTypeClasses [Name]
Does this value contain dictionaries (except of the listed classes).

Since: 0.3

NoUseOf [Name]
Does not contain this value (in terms or patterns)

Since: 0.4.1

CoreOf
Always satisfied, but dumps the value in non-quiet mode.

Since: 0.4.2

Convenience functions

These convenience functions create common test obligations directly.

(===) :: Name -> Name -> Obligation infix 9

Declare two functions to be equal (see EqualTo)

Since: 0.1

(==-) :: Name -> Name -> Obligation infix 9

Declare two functions to be equal, but ignoring type lambdas, type arguments, type casts and hpc ticks (see EqualTo). Note that -fhpc can prevent some optimizations; build without for more reliable analysis.

Since: 0.1.1

(=/=) :: Name -> Name -> Obligation infix 9

Declare two functions to be equal, but expect the test to fail (see EqualTo and expectFail) (This is useful for documentation purposes, or as a TODO list.)

Since: 0.1

(=/-) :: Name -> Name -> Obligation infix 9

Declare two functions to be equal up to types (see (==-)), but expect the test to fail (see expectFail).

Since: 0.4.3.0

(==~) :: Name -> Name -> Obligation infix 9

Declare two functions to be equal as (==-) but also ignoring let bindings ordering (see EqualTo).

Since: 0.5

(=/~) :: Name -> Name -> Obligation infix 9

Declare two functions to be equal up to let binding ordering (see (==~)), but expect the test to fail (see expectFail).

Since: 0.5

hasNoType :: Name -> Name -> Obligation

Declare that in a function’s implementation, the given type does not occur.

More precisely: No locally bound variable (let-bound, lambda-bound or pattern-bound) has a type that contains the given type constructor.

inspect $ fusedFunction `hasNoType` ''[]

Since: 0.1

hasNoGenerics :: Name -> Obligation

Declare that a function’s implementation does not contain any generic types. This is just hasNoType applied to the usual type constructors used in GHC.Generics.

inspect $ hasNoGenerics genericFunction

Since: 0.3

hasNoTypeClasses :: Name -> Obligation

Declare that a function's implementation does not include dictionaries.

More precisely: No locally bound variable (let-bound, lambda-bound or pattern-bound) has a type that contains a type that mentions a type class.

inspect $ hasNoTypeClasses specializedFunction

Since: 0.3

hasNoTypeClassesExcept :: Name -> [Name] -> Obligation

A variant of hasNoTypeClasses, which white-lists some type-classes.

inspect $ fieldLens `hasNoTypeClassesExcept` [''Functor]

Since: 0.3

doesNotUse :: Name -> Name -> Obligation

Declare that a function's implementation does not use the given variable (either in terms or -- if it is a constructor -- in patterns).

inspect $ foo `doesNotUse` 'error

Since: 0.4.1

coreOf :: Name -> Obligation

Dump the Core of the value.

inspect $ coreOf 'foo

This is useful when you need to inspect some values manually.

Since: 0.4.2